Reading all inputs into a generic INI dictionary
Eventually this should be changed a class.  Also, it make be easier to
remove the section nesting when the inputs are being parsed.  Summary
figures are still not being generated but everything else seems to be
working.

Added an additional parameter to allow images to be clipped to the zone
geometry or not.