Major rewrite of Landsat collection generation
User can now set list of Landsat products/bands to compute zonal
statistics for.  This should make it easier to support generating only
missing values or columns in a database.
The collection and image are now being built inside a class in order to
set pass a dictionary of user parameters into the EE functions before
they are constructed.