Switching zonal stats to export entries for all possible SCENE_IDs
This commit will need to get cleaned up to remove some of the DEADBEEF
code that was inserted to get through the meadows data.

Added a “netinfo” export destination that will attempt to write
directly to the output CSV instead of exporting.
Started switching to a “tile” naming scheme instead of “path_row” or
“pr”.  I think this will help when we try to include Sentinel.