Renaming field LOW_ETG_COUNT to ETSTAR_COUNT
This isn't quite the right name either for this value, but it is closer,
since it is actually a function of ET* (not ETg, which the code doesn't
compute).  The problem is that ETSTAR_COUNT implies that it is a count
of pixels with ET* values, instead of a count of low ET* pixels.