Bug fix - Tasumi SR was being scaled by 0.0001 unnecessarily
The USGS_SR data needs to be scaled by 0.0001, but the Tasumi SR is
computed from the TOA which doesn’t need to be scaled.  The scaling is
currently being applied in the generic refl_sur_band_func.  It would
probably make more sense to have dedicated refl_sur_band functions for
Tasumi and USGS_SR, but since USGS_SR will likely become the primary
approach it is easier to just fix Tasumi values in the Tasumi function.