Initial commit of MODIS zonal stats scripts
The guys of the code is setup but the scripts themselves don't work.  I wasn't able to build composite collections of all the possible products, so I instead added an additional for loop to step through each product.  To get this to work I still need to add some code update a subset of columns in the dataframe.

Still need to add cloud masking and return the pixel counts, zenith angle, etc.