More bug fixes and other stuff
Continued working on support for 8 and 16 day products.  Added better support for WKT spatial references.  The calculations are now all default to be done in the MODIS sinusoidal projection, but the cellsizes will change based on the product.  The zenith angle field is now being used to help identify which dates have data and have been processed.  The image wise script seems to be working well.